requirements.txt
amadeus==9.0.0
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
django-ckeditor==6.7.1
django-ckeditor-5==0.2.12
django-js-asset==2.2.0
django-modeltranslation==0.18.11
django-social-share==2.3.0
django-widget-tweaks==1.5.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
outcome==1.3.0
Pillow==9.5.0
psycopg2==2.9.9
pyee==8.2.2
pyppeteer==1.0.2
PySocks==1.7.1
python-decouple==3.8
requests==2.31.0
selenium==4.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.10.0
Unidecode==1.3.8
urllib3==1.26.18
websockets==10.4
whitenoise==6.5.0
wsproto==1.2.0
zipp==3.17.0
